Change log calls to module logging instead of root logger (#3)
* changed CLI tool to click and added entrypoints to setup.cfg

* moved click decorators to main in each module to separate responsibilities

* added extra label and changed to two targets in config and changed log call to be more informative when converting to upload format

* code quality fixes

* code quality darglint

* changed type in click options to click.Path to be able to check if path exists and then convert to pathlib.PAth

* code quality

* removed a mkdir call and added parents flag set to true in downstream mkdir call instead

* moved all click functionality to cli module. removed multiple entrypoints and changed so all calls go through the cli group first

* refactored names of subcommands to be shorter and not clash with imports

* pre-commit checks

* changed wording for copy

* added getlogger call to get module level logging instead of writing everything to root logger

* changed remaining logging calls to logger

* pydocstyle change

* added optional parameters for logging to set logging directory or configure the logger

* started testing how to configure logging using config file

* added support for setting logging config and logging dir

* added check when copying if source file exists and changed so unpacking is on logging config insetup call

* moved getLogger calls from start of module to inside of each function that uses logging

* changed logging configuration to default in config

* moved default config to src package and added manifesto to include it in build, also made config path argument in entrypoint optional

* fixed typo in click arguments for to_uploa_format command

+++ src/wai_data_tools/label_based_file_structure.py
@@ -19,14 +19,16 @@ def copy_files_to_label_based_file_structure(
             the WW01, WW02, ... folders.
         dst_dir: Root directory for label based file structure
     """
-    logging.info("Setting up label based folder structure at %s...", dst_dir)
+    logger = logging.getLogger(__name__)
+
+    logger.info("Setting up label based folder structure at %s...", dst_dir)
 
     labels = file_dataframe["label"].unique()
     for label in labels:
         label_dir = dst_dir / label
         label_dir.mkdir(exist_ok=True, parents=True)
 
-    logging.info("Copying data files to new file structure...")
+    logger.info("Copying data files to new file structure...")
 
     for ind, df_row in tqdm.tqdm(list(file_dataframe.iterrows())):
         filename = df_row["filename"]
@@ -35,4 +37,7 @@ def copy_files_to_label_based_file_structure(
 
         src = src_dir / folder / filename
         dst = dst_dir / label / filename
-        shutil.copy(str(src), str(dst))
+        if src.is_file():
+            shutil.copy(str(src), str(dst))
+        else:
+            logger.debug("Source file %s not found", src.name)
